requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge-score
py7zr
pandas
nltk
tqdm
matplotlib
PyYAML
torch
notebook
boto3
mypy-boto3-s3
python-box
ensure
fastapi
uvicorn
jinja2
jupyter
accelerate -U

pyvi
-e .